third_party/blink/web_tests/external/wpt/css/css-align/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property align-content does not inherit
  assert_equals: expected "last baseline" but got "normal"
Harness: the test ran to completion.
